requirements.txt
Django==2.0
django-bootstrap3==11.0.0
django-ckeditor==5.6.1
pillow==6.0.0
whitenoise==3.2.1
dj-database-url==0.5.0
psycopg2==2.7.5
gunicorn==19.6.0
